requirements.txt
#
# These requirements were autogenerated by pipenv
# To regenerate from the project's Pipfile, run:
#
#    pipenv lock --requirements
#

-i https://pypi.org/simple
asgiref==3.5.0; python_version >= '3.7'
django-environ==0.8.1
django-taggit==2.1.0
django==4.0.2
pillow==9.0.1
psycopg2==2.9.3
sqlparse==0.4.2; python_version >= '3.5'
tzdata==2021.5; sys_platform == 'win32'
